requirements.txt
# AI Bookkeeping System Requirements
# PostgreSQL migration and MCP server dependencies

# Database dependencies
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0

# Data processing (existing dependencies)
pandas>=2.0.0
python-dateutil>=2.8.0

# API dependencies (existing)
requests>=2.28.0

# MCP server dependencies
fastmcp>=0.1.0
pydantic>=2.0.0

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: Database migrations
alembic>=1.12.0

# Optional: Environment management
python-dotenv>=1.0.0

# Optional: Streamlit dashboard (future phase)
streamlit>=1.28.0
plotly>=5.15.0
altair>=5.0.0